% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_package.R
\name{setup_package}
\alias{setup_package}
\title{Setup Personal Package}
\usage{
setup_package(path, packagename, core = NULL)
}
\description{
Setup Personal Package
}
\examples{
if (FALSE) {
  setup_package("..", "mypackage", core = c("dplyr", "glue", "purrr"))
}
}
