% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_core.R
\name{update_core}
\alias{update_core}
\title{Update Core of Personal Package}
\usage{
update_core(path, packagename, core = NULL, append = TRUE)
}
\arguments{
\item{path}{The path in which the package shall be created.
If it exists, it is used. If it does not exist, it is created, provided
that the parent path exists.}

\item{packagename}{The name of the newly generated package. It will be
checked to make sure it meets R package naming conventions.}

\item{core}{A vector or list containing package names that shall be attached
when the newly generated package is loaded. The packages must be installed
on the current system, otherwise an error will be shown.}

\item{append}{If \code{TRUE} the packages of the argument \code{core} will
be appended to the current core and the package version will be increased
on the "patch" level. Otherwise \code{core} will be overwritten and the
package version will be increased on the "minor" level.}
}
\description{
Updates the "core" of a personal package created with personalr. It can
either append another package to the current core or overwrite it
with a new core.
}
