% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalr-package.R
\docType{package}
\name{personalr-package}
\alias{personalr}
\alias{personalr-package}
\title{personalr: Automated Personal Package Setup}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to setup a personal R package that attaches given libraries and exports personal helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrcaseb.github.io/personalr/index.html}
  \item \url{https://github.com/mrcaseb/personalr}
  \item Report bugs at \url{https://github.com/mrcaseb/personalr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{internal}
